:sparkles: v0.2.0 Pre-Release
🏷️ Begin dependency version tracking

✨ Add snipe status handling

:fire: Remove snipe_id

:fire: Re: Remove snipe_id

:arrow_up: Upgrade moonbois_core -> v0.2.0

:heavy_plus_sign: Add spinoff

:arrow_up: Bump moonbois_core -> v0.2.1

:arrow_up: Bump dialoguer -> v0.1.1

:stethoscope: Rename package -> moonbois_cli

:stethoscope: Fix stack overflow issue

:stethoscope: Fix stack overflow issue -> main.rs

:sparkles: Move Loader from dialogue | :bug:  detect_enter fixes/reworks

:sparkles: Add Bumpbot

:see_no_evil: Update .gitignore

:sparkles: Add release.iss

:fire: Remove macos/linux support

:fire: Remove unused vars

:sparkles: Use inno for packaging

:wrench: File mode change

:wrench: Set default-run to moonbois_cli
