+++ .gitignore
@@ -1,1 +1,4 @@
 target
+.cargo
+output
+Cargo.lock
