+++ src/bin/detect_enter.rs
@@ -1,0 +1,4 @@
+fn main() {
+    let mut input = String::new();
+    std::io::stdin().read_line(&mut input).expect("Failed to read input");
+}
