+++ src/main.rs
@@ -3,30 +3,18 @@ use std::sync::Arc;
 use std::time::Duration;
 use console::style;
 use dialoguer::theme::ColorfulTheme;
+use crate::dialogue::loader::LoaderError;
 use dialoguer::Select;
-use handlers::Buy;
-use handlers::CancelSnipe;
-use handlers::CreateProject;
-use handlers::CreateSnipe;
-use handlers::DeleteProject;
-use handlers::DeleteWallet;
-use handlers::Deposit;
-use handlers::Export;
+
+use handlers::bumps::BumpMenu;
+use handlers::bumps::StartBumps;
+use handlers::bumps::StopBumps;
 use handlers::Handler;
-use handlers::ImportWallet;
-use handlers::RecoverSol;
-use handlers::SelectProject;
-use handlers::Login;
-use handlers::MainMenu;
-use handlers::ProjectMenu;
-use handlers::Sell;
-use handlers::SendSOL;
-use handlers::Signup;
-use handlers::WalletMenu;
-use handlers::Withdraw;
 use moonbois_core::rpc::MoonboisClient;
 use moonbois_core::rpc::MoonboisClientError;
 use moonbois_core::ProjectDTO;
+use moonbois_core::PendingSnipeError;
+use moonbois_core::PumpfunBumpStatus;
 use moonbois_core::UserDTO;
 use solana_sdk::native_token::LAMPORTS_PER_SOL;
 use solana_sdk::pubkey::ParsePubkeyError;
@@ -35,7 +23,15 @@ use tokio::task::JoinHandle;
 use tokio::time::sleep_until;
 use tokio::time::Instant;
 
+use handlers::auth::*;
+use handlers::trade::*;
+use handlers::wallet::*;
+use handlers::snipe::*;
+use handlers::project::*;
+use handlers::main::*;
+
 pub mod handlers;
+pub mod dialogue;
 
 static BANNER: &str = r#"
  _____ _____ _____ _____ _____ _____ _____ _____ 
@@ -46,23 +42,28 @@ static BANNER: &str = r#"
 
 pub struct ActiveProject(pub Option<i32>);
 pub struct ActiveUser(pub Option<UserDTO>);
+pub struct BumpStatus(pub Option<PumpfunBumpStatus>);
 
 pub struct AppData {
     pub rpc_client: RwLock<MoonboisClient>,
     pub user: RwLock<ActiveUser>,
     pub projects: RwLock<HashMap<i32, ProjectDTO>>,
-    pub active_project: RwLock<ActiveProject>
+    pub active_project: RwLock<ActiveProject>,
+    pub bump_status: RwLock<BumpStatus>
 }
 
 pub enum Menu {
     Main(MainMenu),
     Login(Login),
+    Bump(BumpMenu),
     Signup(Signup),
     Wallet(WalletMenu),
     Send(SendSOL),
     ImportWallet(ImportWallet),
     DeleteWallet(DeleteWallet),
     Buy(Buy),
+    StartBumps(StartBumps),
+    StopBumps(StopBumps),
     DeleteProject(DeleteProject),
     CancelSnipe(CancelSnipe),
     CreateSnipe(CreateSnipe),
@@ -76,7 +77,7 @@ pub enum Menu {
     Export(Export)
 }
 impl Handler for Menu {
-    async fn handle(self, app_data: &Arc<AppData>) -> Result<Option<Menu>, (Menu, AppError)> {
+    async fn handle(&self, app_data: &Arc<AppData>) -> Result<Option<Menu>, (Menu, AppError)> {
         match self {
             Menu::Main(handler) => handler.handle(app_data).await,
             Menu::Login(handler) => handler.handle(app_data).await,
@@ -85,6 +86,9 @@ impl Handler for Menu {
             Menu::Wallet(handler) => handler.handle(app_data).await,
             Menu::ProjectMenu(handler) => handler.handle(app_data).await,
             Menu::SelectProject(handler) => handler.handle(app_data).await,
+            Menu::StartBumps(handler) => handler.handle(app_data).await,
+            Menu::StopBumps(handler) => handler.handle(app_data).await,
+            Menu::Bump(handler) => handler.handle(app_data).await,
             Menu::Buy(handler) => handler.handle(app_data).await, 
             Menu::DeleteWallet(handler) => handler.handle(app_data).await, 
             Menu::ImportWallet(handler) => handler.handle(app_data).await, 
@@ -155,69 +159,112 @@ impl App {
                         }
                     }
 
+                    let rpc_client = app_data_arc.rpc_client.read().await;
+                    let bump_status = match rpc_client.get_bumps_status().await {
+                        Ok(bump_status) => Some(bump_status),
+                        Err(MoonboisClientError::NotFound) => None,
+                        Err(err) => return Err(err.to_string())
+                    };
+                    let mut bump_status_write = app_data_arc.bump_status.write().await;
+                    bump_status_write.0 = bump_status;
+
+                    drop(rpc_client);
+
                     sleep_until(Instant::now() + Duration::from_millis(500)).await;
                 }
             })
         }
     }
-    pub async fn run(self, menu: Menu) {
-        std::process::Command::new("clear").status().unwrap();
-        println!("{}", style(BANNER).bold());
-        
-        if let Some(user) = &self.app_data.user.read().await.0 {
-            let sniper_balance = user.wallets.iter().map(|(_, y)| y.sol_balance).reduce(|x, y| x + y).unwrap() as f64 / LAMPORTS_PER_SOL as f64;
-            let user_balance = user.sol_balance as f64 / LAMPORTS_PER_SOL as f64;
-            println!(
-                "fee_payer: {}\nfee_payer_balance: {}\nsniper_sol_balance: {}",
-                user.public_key, 
-                format!("{} {}", user_balance, style("SOL").cyan()),
-                format!("{} {}", sniper_balance, style("SOL").cyan()),
-            );
+    pub async fn run(self, mut current_menu: Menu) {
+        loop {
+            std::process::Command::new("clear").status().unwrap();
+            println!("{}", style(BANNER).bold());
+            
+            if let Some(user) = &self.app_data.user.read().await.0 {
+                let sniper_balance = user.wallets.iter().map(|(_, y)| y.sol_balance).reduce(|x, y| x + y).unwrap() as f64 / LAMPORTS_PER_SOL as f64;
+                let user_balance = user.sol_balance as f64 / LAMPORTS_PER_SOL as f64;
+                println!(
+                    "fee_payer: {}\nfee_payer_balance: {}\nsniper_sol_balance: {}",
+                    user.public_key, 
+                    format!("{} {}", user_balance, style("SOL").cyan()),
+                    format!("{} {}", sniper_balance, style("SOL").cyan()),
+                );
+                if let Some(active_project) = &self.app_data.active_project.read().await.0 {
+                    if let Some(project) = &self.app_data.projects.read().await.get(active_project) {
+                        let sniper_token_balance = user.wallets.iter().filter_map(|(_, x)| x.token_balance).reduce(|a, b| a + b).unwrap_or(0) as f64 / 10f64.powf(6f64);
+                        println!(
+                            "snipe_token_balance: {} {}",
+                            format!("{}", sniper_token_balance),
+                            format!("{}", style(project.name.to_uppercase()).magenta())
+                        )
+                    }
+                }
+            }
+    
             if let Some(active_project) = &self.app_data.active_project.read().await.0 {
+                if let Some(bump_status) = &self.app_data.bump_status.read().await.0 {
+                    match bump_status {
+                        PumpfunBumpStatus::Failed(reason) => println!("bump_status: {}", reason),
+                        PumpfunBumpStatus::Pending => println!("bump_status: pending"),
+                        PumpfunBumpStatus::Running => println!("bump_status: running"),
+                    }                
+                } else {
+                    println!("bump_status: not started");
+                }
                 if let Some(project) = &self.app_data.projects.read().await.get(active_project) {
-                    let sniper_token_balance = user.wallets.iter().filter_map(|(_, x)| x.token_balance).reduce(|a, b| a + b).unwrap_or(0) as f64 / 10f64.powf(6f64);
                     println!(
-                        "snipe_token_balance: {} {}",
-                        format!("{}", sniper_token_balance),
-                        format!("{}", style(project.name.to_uppercase()).magenta())
+                        "mint_id: {}\ndeployer: {}",
+                        project.pumpfun.mint_id,
+                        project.deployer
                     )
                 }
             }
-        }
-
-        if let Some(active_project) = &self.app_data.active_project.read().await.0 {
-            if let Some(project) = &self.app_data.projects.read().await.get(active_project) {
-                println!(
-                    "mint_id: {}\ndeployer: {}",
-                    project.pumpfun.mint_id,
-                    project.deployer
-                )
+    
+            println!("");
+    
+            if self.socket_handle.is_finished() {
+                let error_message = if let Err(err) = self.socket_handle.await.unwrap() {
+                    err
+                } else {
+                    "".to_string()
+                };
+    
+                println!("\n{}\n{}", style("Websocket connection failed ⚠️").yellow(), style(error_message).dim());
+                Select::with_theme(&ColorfulTheme::default())
+                    .items(&vec!["Back"])
+                    .default(0)
+                    .interact()
+                    .unwrap();
+                
+                return;
             }
-        }
+    
+            // TODO - find better way to handle results
+            match current_menu.handle(&self.app_data).await {
+                Ok(Some(result)) => {
+                    current_menu = result;
+                }
+                Err((menu, AppError::LoaderError(err))) => {
+                    println!("{}\n  - {}", style("Load failed ⚠️").yellow(), style(err).dim());
+                    Select::with_theme(&ColorfulTheme::default())
+                        .items(&vec!["Back"])
+                        .default(0)
+                        .interact()
+                        .unwrap();
 
-        println!("");
+                    current_menu = menu;
+                }
+                Err((menu, AppError::PendingSnipeError(err))) => {
+                    println!("{}\n  - {}", style("Pending snipe failed ⚠️").yellow(), style(err).dim());
+                    Select::with_theme(&ColorfulTheme::default())
+                        .items(&vec!["Back"])
+                        .default(0)
+                        .interact()
+                        .unwrap();
 
-        if self.socket_handle.is_finished() {
-            let error_message = if let Err(err) = self.socket_handle.await.unwrap() {
-                err
-            } else {
-                "".to_string()
-            };
-
-            println!("\n{}\n{}", style("Websocket connection failed ⚠️").yellow(), style(error_message).dim());
-            Select::with_theme(&ColorfulTheme::default())
-                .items(&vec!["Back"])
-                .default(0)
-                .interact()
-                .unwrap();
-            
-            return;
-        }
-
-        let handler_response = menu.handle(&self.app_data).await;
-        if let Err((menu, err)) = handler_response {
-            match err {
-                AppError::MoonboisClientError(MoonboisClientError::UnhandledServerError(err)) => {
+                    current_menu = menu;
+                }
+                Err((menu, AppError::MoonboisClientError(MoonboisClientError::UnhandledServerError(err)))) => {
                     println!("{}\n  - {}", style("Unhandled server error occured ⚠️").yellow(), style(err).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -225,9 +272,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::MoonboisClientError(MoonboisClientError::NotFound) => {
+                Err((menu, AppError::MoonboisClientError(MoonboisClientError::NotFound))) => {
                     println!("{}", style("Requested resource was not found ⚠️").yellow());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -235,9 +282,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::MoonboisClientError(MoonboisClientError::InvalidUri(err)) => {
+                Err((menu, AppError::MoonboisClientError(MoonboisClientError::InvalidUri(err)))) => {
                     println!("{}\n  - {}", style("Invalid URI ⚠️").yellow(), style(err.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -245,9 +292,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::MoonboisClientError(MoonboisClientError::JsonError(err)) => {
+                Err((menu, AppError::MoonboisClientError(MoonboisClientError::JsonError(err)))) => {
                     println!("{}\n  - {}", style("JSON error ⚠️").yellow(), style(err.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -255,9 +302,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::MoonboisClientError(MoonboisClientError::MissingJWT) => {
+                Err((_menu, AppError::MoonboisClientError(MoonboisClientError::MissingJWT))) => {
                     println!("{}", style("Authorization failed ⚠️").yellow());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -265,9 +312,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, Menu::Login(Login))).await;
+                    current_menu = Menu::Login(Login);
                 }
-                AppError::MoonboisClientError(MoonboisClientError::ParseError(err)) => {
+                Err((menu, AppError::MoonboisClientError(MoonboisClientError::ParseError(err)))) => {
                     println!("{}\n  - {}", style("Parse error occured ⚠️").yellow(), style(err.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -275,9 +322,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::MoonboisClientError(MoonboisClientError::ReqwestError(err)) => {
+                Err((menu, AppError::MoonboisClientError(MoonboisClientError::ReqwestError(err)))) => {
                     println!("{}\n  - {}", style("An error occured ⚠️").yellow(), err);
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -285,9 +332,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::MoonboisClientError(MoonboisClientError::NotAccepted) => {
+                Err((menu, AppError::MoonboisClientError(err))) => {
                     println!("{}\n  - {}", style("Not accepted ⚠️").yellow(), style(err.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -295,9 +342,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::ParsePubkeyError(err) => {
+                Err((menu, AppError::ParsePubkeyError(err))) => {
                     println!("{}\n  - {}", style("Invalid pubkey ⚠️").yellow(), style(err.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -305,9 +352,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::DialogueError(err) => {
+                Err((menu, AppError::DialogueError(err))) => {
                     println!("{}\n  - {}", style("Dialogue error occured ⚠️").yellow(), style(err.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -315,9 +362,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::ProjectNotFound => {
+                Err((menu, AppError::ProjectNotFound)) => {
                     println!("{}\n", style("Project not found ⚠️").yellow());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -325,9 +372,9 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::Unhandled(err) => {
+                Err((menu, AppError::Unhandled(err))) => {
                     println!("{}\n - {}", style("Unhandled error occured ⚠️").yellow(), err);
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
@@ -335,21 +382,18 @@ impl App {
                         .interact()
                         .unwrap();
 
-                        let _ = Box::pin(App::run(self, menu)).await;
+                    current_menu = menu;
                 }
-                AppError::UserNotFound => {
-                    println!("{}\n  - {}", style("Unable to find user ⚠️").yellow(), style(err.to_string()).dim());
+                Err((_menu, AppError::UserNotFound)) => {
+                    println!("{}\n  - {}", style("Unable to find user ⚠️").yellow(), style(AppError::UserNotFound.to_string()).dim());
                     Select::with_theme(&ColorfulTheme::default())
                         .items(&vec!["Back"])
                         .default(0)
                         .interact()
                         .unwrap();
-
-                        let _ = Box::pin(App::run(self, menu)).await;
                 }
+                Ok(None) => break
             }
-        } else if let Ok(Some(result)) = handler_response {
-            let _ = Box::pin(App::run(self, result)).await;
         }
     }
 }
@@ -358,6 +402,7 @@ impl App {
 pub async fn main() {
     let app_data = Arc::new(AppData {
         active_project: RwLock::new(ActiveProject(None)),
+        bump_status: RwLock::new(BumpStatus(None)),
         projects: RwLock::new(HashMap::new()),
         rpc_client: RwLock::new(MoonboisClient::new()),
         user: RwLock::new(ActiveUser(None))
@@ -371,10 +416,14 @@ pub async fn main() {
 pub enum AppError {
     #[error("Moonbois client error: {0}")]
     MoonboisClientError(#[from] MoonboisClientError),
+    #[error("Pending snipe error: {0}")]
+    PendingSnipeError(#[from] PendingSnipeError),
     #[error("Dialogue error: {0}")]
     DialogueError(#[from] dialoguer::Error),
-    #[error("Parse pubkeyu error: {0}")]
+    #[error("Parse pubkey error: {0}")]
     ParsePubkeyError(#[from] ParsePubkeyError),
+    #[error("Loader error: {0}")]
+    LoaderError(#[from] LoaderError),
     #[error("Project not found")]
     ProjectNotFound,
     #[error("User not found")]
